refact navbar part of the Header component
