+++ frontend-opal/components/Header/Header.js
@@ -1,26 +1,16 @@
 import React from 'react';
-import { Button, View, TextInput, Image } from 'react-native';
+import { View, TextInput, Image } from 'react-native';
 import styles from './styles';
 import ListIcon from '../../assets/list-icon.png';
 import SearchIcon from '../../assets/search-button.png';
 
+import NavBar from './NavBar';
+
 const Header = () => {
-  const narTitles = [
-    { id: 1, key: 'PORTFOLI' },
+  const navTitles = [
+    { id: 1, key: 'PORTFOLIO' },
     { id: 2, key: 'WATCHLIST' }
   ];
-
-  const handlePress = key => key === 'PORTFOLIO';
-
-  const renderItems = titles =>
-    titles.map(title => (
-      <View
-        key={title.id}
-        style={[styles.narTitle, handlePress(title.key) && styles.buttonPressed]}
-      >
-        <Button title={title.key} color="#e17055" />
-      </View>
-    ));
 
   return (
     <View>
@@ -30,7 +20,7 @@ const Header = () => {
         <Image style={styles.searchButton} source={SearchIcon} />
       </View>
 
-      <View style={styles.navBars}>{renderItems(narTitles)}</View>
+      <NavBar navTitles={navTitles} />
     </View>
   );
 };
